separated mainmenu() method
